Prototype
The prototype includes the solr implementation with example docs indexed
and stored acording to custom 'scheme.xml'. I edited the default search
handler in 'solrconfig.xml' as well to work with the coustom scheme.
I'm working out of the 'SearchEncounterIDECommands' directory under the
example configuration.
Added the original manuals and the example
fetxtcmdref-mini_example-solr.xml which I created concisting of a few
commands from the manuals. That example xml was indexd in this solr
instance for searching.